requirements.txt
###### Requirements without Version Specifiers ######
Finance
Streamline
Pandas
Datetime
